OpenAPI & UI: Improve Implementation
